qwerty

+++ core/main-service/src/main/java/ru/practicum/main/managedlocation/controller/AdminManagedLocationController.java
@@ -5,6 +5,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import ru.practicum.main.exception.ValidationException;
 import ru.practicum.main.managedlocation.dto.ManagedLocationDto;
 import ru.practicum.main.managedlocation.dto.NewManagedLocationDto;
 import ru.practicum.main.managedlocation.dto.UpdateManagedLocationDto;
@@ -23,7 +24,6 @@ public class AdminManagedLocationController {
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public ManagedLocationDto create(@Valid @RequestBody NewManagedLocationDto dto) {
-        log.info("POST /admin/locations");
         return service.create(dto);
     }
 
@@ -31,21 +31,27 @@ public class AdminManagedLocationController {
     @ResponseStatus(HttpStatus.OK)
     public ManagedLocationDto update(@PathVariable Long locationId,
                                      @RequestBody UpdateManagedLocationDto dto) {
-        log.info("PATCH /admin/locations/{}", locationId);
+        if (locationId <= 0) {
+            throw new ValidationException("locationId must be positive");
+        }
         return service.update(locationId, dto);
     }
 
     @DeleteMapping("/{locationId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable Long locationId) {
-        log.info("DELETE /admin/locations/{}", locationId);
+        if (locationId <= 0) {
+            throw new ValidationException("locationId must be positive");
+        }
         service.delete(locationId);
     }
 
     @GetMapping("/{locationId}")
     @ResponseStatus(HttpStatus.OK)
     public ManagedLocationDto getById(@PathVariable Long locationId) {
-        log.info("GET /admin/locations/{}", locationId);
+        if (locationId <= 0) {
+            throw new ValidationException("locationId must be positive");
+        }
         return service.getById(locationId);
     }
 
@@ -53,7 +59,12 @@ public class AdminManagedLocationController {
     @ResponseStatus(HttpStatus.OK)
     public List<ManagedLocationDto> getAll(@RequestParam(defaultValue = "0") int from,
                                            @RequestParam(defaultValue = "10") int size) {
-        log.info("GET /admin/locations?from={}&size={}", from, size);
+        if (size <= 0) {
+            throw new ValidationException("size must be positive");
+        }
+        if (from < 0) {
+            throw new ValidationException("from must be >= 0");
+        }
         return service.getAll(from, size);
     }
 }
